moviezzz/requirements.txt
Django==4.2.16
djangorestframework==3.14.0
gunicorn==23.0.0
psycopg2-binary==2.9.9
whitenoise==6.7.0
django-cors-headers==4.3.1
